refactor: refactored brands components and pages
