+++ web/src/constants/brand.constants.ts
@@ -3,3 +3,13 @@ export const BRAND_SORT_OPTIONS = [
   { value: 'name', label: 'Name (A–Z)' },
   { value: 'events_count', label: 'Most events' },
 ];
+
+export const ITEMS_PER_PAGE = 6;
+
+export const ACTIVE_STATUSES = new Set([
+  'published',
+  'published_unverified',
+  'published_on_review',
+]);
+
+export const PENDING_STATUSES = ['draft_on_review', 'published_on_review'];
